+++ src/app.tsx
@@ -1,47 +1,20 @@
-import { Providers, ProviderType } from 'api/providers';
-import { getProvidersQuery } from 'api/queries/providersQuery';
-import { AxiosError } from 'axios';
 import { I18nProvider } from 'components/i18n';
-import { InactiveSources } from 'components/sources/InactiveSources/InactiveSources';
 import Maintenance from 'pages/state/maintenance';
-import NotAuthorized from 'pages/state/notAuthorized';
-import NotAvailable from 'pages/state/notAvailable';
 import React from 'react';
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router';
 import { withRouter } from 'react-router-dom';
 import { compose } from 'redux';
-import { createMapStateToProps, FetchStatus } from 'store/common';
-import {
-  awsProvidersQuery,
-  azureProvidersQuery,
-  ocpProvidersQuery,
-  providersActions,
-  providersSelectors,
-} from 'store/providers';
+import { createMapStateToProps } from 'store/common';
 
 import { Routes, routes } from './routes';
 
 export interface AppOwnProps extends RouteComponentProps<void> {}
 
-interface AppStateProps {
-  awsProviders: Providers;
-  awsProvidersError: AxiosError;
-  awsProvidersFetchStatus: FetchStatus;
-  awsProvidersQueryString: string;
-  azureProviders: Providers;
-  azureProvidersError: AxiosError;
-  azureProvidersFetchStatus: FetchStatus;
-  azureProvidersQueryString: string;
-  ocpProviders: Providers;
-  ocpProvidersError: AxiosError;
-  ocpProvidersFetchStatus: FetchStatus;
-  ocpProvidersQueryString: string;
-}
+interface AppStateProps {}
 
 interface AppDispatchProps {
   history: any;
-  fetchProviders: typeof providersActions.fetchProviders;
 }
 
 interface AppState {
@@ -59,15 +32,7 @@ export class App extends React.Component<AppProps, AppState> {
   public state: AppState = { locale: 'en', maintenanceMode: false };
 
   public componentDidMount() {
-    const {
-      awsProviders,
-      awsProvidersFetchStatus,
-      azureProviders,
-      azureProvidersFetchStatus,
-      history,
-      ocpProviders,
-      ocpProvidersFetchStatus,
-    } = this.props;
+    const { history } = this.props;
 
     insights.chrome.init();
     insights.chrome.identifyApp('cost-management');
@@ -78,43 +43,11 @@ export class App extends React.Component<AppProps, AppState> {
         history.push(currRoute.path);
       }
     });
-
-    // Must fetch for all pages where InactiveSources may be shown
-    if (!awsProviders && awsProvidersFetchStatus !== FetchStatus.inProgress) {
-      this.fetchAwsProviders();
-    }
-    if (!azureProviders && azureProvidersFetchStatus !== FetchStatus.inProgress) {
-      this.fetchAzureProviders();
-    }
-    if (!ocpProviders && ocpProvidersFetchStatus !== FetchStatus.inProgress) {
-      this.fetchOcpProviders();
-    }
   }
 
   public componentDidUpdate(prevProps: AppProps) {
-    const {
-      awsProviders,
-      awsProvidersError,
-      awsProvidersFetchStatus,
-      azureProviders,
-      azureProvidersError,
-      azureProvidersFetchStatus,
-      location,
-      ocpProviders,
-      ocpProvidersError,
-      ocpProvidersFetchStatus,
-    } = this.props;
+    const { location } = this.props;
 
-    // Must fetch for all pages where InactiveSources may be shown
-    if (!awsProviders && awsProvidersFetchStatus !== FetchStatus.inProgress && !awsProvidersError) {
-      this.fetchAwsProviders();
-    }
-    if (!azureProviders && azureProvidersFetchStatus !== FetchStatus.inProgress && !azureProvidersError) {
-      this.fetchAzureProviders();
-    }
-    if (!ocpProviders && ocpProvidersFetchStatus !== FetchStatus.inProgress && !ocpProvidersError) {
-      this.fetchOcpProviders();
-    }
     if (location && location.pathname !== prevProps.location.pathname) {
       window.scrollTo(0, 0);
     }
@@ -124,101 +57,19 @@ export class App extends React.Component<AppProps, AppState> {
     this.appNav();
   }
 
-  private fetchAwsProviders = () => {
-    const { awsProvidersQueryString, fetchProviders } = this.props;
-    fetchProviders(ProviderType.aws, awsProvidersQueryString);
-  };
-
-  private fetchAzureProviders = () => {
-    const { azureProvidersQueryString, fetchProviders } = this.props;
-    fetchProviders(ProviderType.azure, azureProvidersQueryString);
-  };
-
-  private fetchOcpProviders = () => {
-    const { fetchProviders, ocpProvidersQueryString } = this.props;
-    fetchProviders(ProviderType.ocp, ocpProvidersQueryString);
-  };
-
   public render() {
-    const { awsProvidersError, azureProvidersError, ocpProvidersError } = this.props;
     const { maintenanceMode } = this.state;
-    let route = <Routes />;
+    const route = maintenanceMode ? <Maintenance /> : <Routes />;
 
-    if (maintenanceMode) {
-      route = <Maintenance />;
-    } else {
-      // Note that the sources API will return 403 when the user doesn't have entitlements
-      const error = awsProvidersError || azureProvidersError || ocpProvidersError;
-
-      if (error) {
-        if (error.response && (error.response.status === 401 || error.response.status === 403)) {
-          route = <NotAuthorized />;
-        } else {
-          route = <NotAvailable />;
-        }
-      }
-    }
-    return (
-      <I18nProvider locale={this.state.locale}>
-        <InactiveSources />
-        {route}
-      </I18nProvider>
-    );
+    return <I18nProvider locale={this.state.locale}>{route}</I18nProvider>;
   }
 }
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const mapStateToProps = createMapStateToProps<AppOwnProps, AppStateProps>((state, props) => {
-  const awsProvidersQueryString = getProvidersQuery(awsProvidersQuery);
-  const awsProviders = providersSelectors.selectProviders(state, ProviderType.aws, awsProvidersQueryString);
-  const awsProvidersError = providersSelectors.selectProvidersError(state, ProviderType.aws, awsProvidersQueryString);
-  const awsProvidersFetchStatus = providersSelectors.selectProvidersFetchStatus(
-    state,
-    ProviderType.aws,
-    awsProvidersQueryString
-  );
-
-  const azureProvidersQueryString = getProvidersQuery(azureProvidersQuery);
-  const azureProviders = providersSelectors.selectProviders(state, ProviderType.azure, azureProvidersQueryString);
-  const azureProvidersError = providersSelectors.selectProvidersError(
-    state,
-    ProviderType.azure,
-    azureProvidersQueryString
-  );
-  const azureProvidersFetchStatus = providersSelectors.selectProvidersFetchStatus(
-    state,
-    ProviderType.azure,
-    azureProvidersQueryString
-  );
-
-  const ocpProvidersQueryString = getProvidersQuery(ocpProvidersQuery);
-  const ocpProviders = providersSelectors.selectProviders(state, ProviderType.ocp, ocpProvidersQueryString);
-  const ocpProvidersError = providersSelectors.selectProvidersError(state, ProviderType.ocp, ocpProvidersQueryString);
-  const ocpProvidersFetchStatus = providersSelectors.selectProvidersFetchStatus(
-    state,
-    ProviderType.ocp,
-    ocpProvidersQueryString
-  );
-
-  return {
-    awsProviders,
-    awsProvidersError,
-    awsProvidersFetchStatus,
-    awsProvidersQueryString,
-    azureProviders,
-    azureProvidersError,
-    azureProvidersFetchStatus,
-    azureProvidersQueryString,
-    ocpProviders,
-    ocpProvidersError,
-    ocpProvidersFetchStatus,
-    ocpProvidersQueryString,
-  };
+  return {};
 });
 
-const mapDispatchToProps: AppDispatchProps = {
-  history,
-  fetchProviders: providersActions.fetchProviders,
-};
+const mapDispatchToProps: AppDispatchProps = { history };
 
 export default compose(withRouter, connect(mapStateToProps, mapDispatchToProps))(App);
