+++ src/components/sources/inactiveSources/index.ts
@@ -1,0 +1,3 @@
+import { InactiveSources } from './inactiveSources';
+
+export default InactiveSources;
